pyproject.toml
[project]
name = "cartie"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cohere>=5.15.0",
    "mysql-connector-python>=9.3.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
